+++ chainer/optimizer.py
@@ -1,5 +1,6 @@
 import collections
 import copy
+import sys
 import warnings
 
 import numpy
@@ -7,6 +8,7 @@ import six
 
 from chainer import cuda
 from chainer import link as link_module
+from chainer import optimizer_hooks
 from chainer import serializer as serializer_module
 from chainer import variable
 
@@ -19,16 +21,6 @@ def _sum_sqnorm(arr):
             s = x.dot(x)
             sq_sum[int(dev)] += s
     return sum([float(i) for i in six.itervalues(sq_sum)])
-
-
-def exponential_decay_noise(xp, shape, dtype, hook, opt):
-    """Time-dependent annealed Gaussian noise function from the paper:
-
-    `Adding Gradient Noise Improves Learning for Very Deep Networks
-    <https://arxiv.org/pdf/1511.06807>`_.
-    """
-    std = numpy.sqrt(hook.eta / numpy.power(1 + opt.t, 0.55))
-    return xp.random.normal(0, std, shape).astype(dtype)
 
 
 class Hyperparameter(object):
@@ -603,171 +595,42 @@ class HyperparameterProxy(object):
         setattr(obj.hyperparam, self._attr_name, value)
 
 
-class WeightDecay(object):
+def deprecation_warning_wrappter(mod, deprecated):
 
-    """Optimizer/UpdateRule hook function for weight decay regularization.
+    deprecated = set(deprecated)
 
-    This hook function adds a scaled parameter to the corresponding gradient.
-    It can be used as a regularization.
+    class Wrapper(object):
 
-    Args:
-        rate (float): Coefficient for the weight decay.
+        def __getattr__(self, attr):
+            if attr in deprecated:
+                warnings.warn(
+                    "chainer.optimizer.{0} is deprecated since v3. "
+                    "Use chainer.optimizer_hooks.{0} instead.".format(attr),
+                    DeprecationWarning)
 
-    Attributes:
-        rate (float): Coefficient for the weight decay.
+            return getattr(mod, attr)
 
-    """
-    name = 'WeightDecay'
-    call_for_each_param = True
-
-    def __init__(self, rate):
-        self.rate = rate
-
-    def __call__(self, rule, param):
-        p, g = param.data, param.grad
-        with cuda.get_device_from_array(p) as dev:
-            if int(dev) == -1:
-                g += self.rate * p
-            else:
-                kernel = cuda.elementwise(
-                    'T p, T decay', 'T g', 'g += decay * p', 'weight_decay')
-                kernel(p, self.rate, g)
+        def __setattr__(self, attr, value):
+            if attr in deprecated:
+                warnings.warn(
+                    "chainer.optimizer.{0} is deprecated since v3. "
+                    "Use chainer.optimizer_hooks.{0} instead".format(attr),
+                    DeprecationWarning)
+            return setattr(mod, attr, value)
+    return Wrapper()
 
 
-class Lasso(object):
-    """Optimizer/UpdateRule hook function for Lasso regularization.
+#  Backward compatibility
+GradientClipping = optimizer_hooks.GradientClipping
+GradientHardClipping = optimizer_hooks.GradientHardClipping
+GradientNoise = optimizer_hooks.GradientNoise
+Lasso = optimizer_hooks.Lasso
+WeightDecay = optimizer_hooks.WeightDecay
 
-    This hook function adds a scaled parameter to the sign of each weight.
-    It can be used as a regularization.
-
-    Args:
-        rate (float): Coefficient for the weight decay.
-
-    Attributes:
-        rate (float): Coefficient for the weight decay.
-
-    """
-    name = 'Lasso'
-    call_for_each_param = True
-
-    def __init__(self, rate):
-        self.rate = rate
-
-    def __call__(self, rule, param):
-        p, g = param.data, param.grad
-        xp = cuda.get_array_module(p)
-        with cuda.get_device_from_array(p) as dev:
-            sign = xp.sign(p)
-            if int(dev) == -1:
-                g += self.rate * sign
-            else:
-                kernel = cuda.elementwise(
-                    'T s, T decay', 'T g', 'g += decay * s', 'lasso')
-                kernel(sign, self.rate, g)
-
-
-class GradientClipping(object):
-    """Optimizer hook function for gradient clipping.
-
-    This hook function scales all gradient arrays to fit to the defined L2 norm
-    threshold.
-
-    Args:
-        threshold (float): L2 norm threshold.
-
-    Attributes:
-        threshold (float): L2 norm threshold of gradient norm.
-
-    """
-    name = 'GradientClipping'
-
-    def __init__(self, threshold):
-        self.threshold = threshold
-
-    def __call__(self, opt):
-        norm = numpy.sqrt(_sum_sqnorm(
-            [p.grad for p in opt.target.params(False)]))
-        rate = self.threshold / norm
-        if rate < 1:
-            for param in opt.target.params(False):
-                grad = param.grad
-                with cuda.get_device_from_array(grad):
-                    grad *= rate
-
-
-class GradientNoise(object):
-    """Optimizer/UpdateRule hook function for adding gradient noise.
-
-    This hook function simply adds noise generated by the ``noise_func``
-    to the gradient. By default it adds time-dependent annealed Gaussian
-    noise to the gradient at every training step:
-
-    .. math::
-
-        g_t \\leftarrow g_t + N(0, \\sigma_t^2)
-
-    where
-
-    .. math::
-
-        \\sigma_t^2 = \\frac{\\eta}{(1+t)^\\gamma}
-
-    with :math:`\\eta` selected from {0.01, 0.3, 1.0} and
-    :math:`\\gamma = 0.55`.
-
-    Args:
-        eta (float): Parameter that defines the scale of the noise, which for
-            the default noise function is recommended to be either 0.01, 0.3
-            or 1.0.
-        noise_func (function): Noise generating function which by default
-            is given by `Adding Gradient Noise Improves Learning for Very Deep\
-            Networks <https://arxiv.org/pdf/1511.06807>`_.
-    """
-    name = 'GradientNoise'
-    call_for_each_param = True
-
-    def __init__(self, eta, noise_func=exponential_decay_noise):
-        self.eta = eta
-        self.noise_func = noise_func
-
-    def __call__(self, rule, param):
-        g = param.grad
-        xp = cuda.get_array_module(g)
-        with cuda.get_device_from_array(g) as dev:
-            noise = self.noise_func(xp, g.shape, g.dtype, self, rule)
-            if int(dev) == -1:
-                g += noise
-            else:
-                kernel = cuda.elementwise(
-                    'T noise', 'T g', 'g += noise', 'gradient_noise')
-                kernel(noise, g)
-
-
-class GradientHardClipping(object):
-
-    """Optimizer/UpdateRule hook function for gradient clipping.
-
-    This hook function clips all gradient arrays to be within a lower and upper
-    bound.
-
-    Args:
-        lower_bound (float): The lower bound of the gradient value.
-        upper_bound (float): The upper bound of the gradient value.
-
-    Attributes:
-        lower_bound (float): The lower bound of the gradient value.
-        upper_bound (float): The upper bound of the gradient value.
-
-    """
-    name = 'GradientHardClipping'
-    call_for_each_param = True
-
-    def __init__(self, lower_bound, upper_bound):
-        self.lower_bound = lower_bound
-        self.upper_bound = upper_bound
-
-    def __call__(self, rule, param):
-        grad = param.grad
-        xp = cuda.get_array_module(grad)
-        with cuda.get_device_from_array(grad):
-            xp.clip(grad, self.lower_bound, self.upper_bound, out=grad)
+sys.modules[__name__] = deprecation_warning_wrapper(
+    sys.modules[__name__],
+    deprecated=['GradientClipping',
+                'GradientHardClipping',
+                'GradientNoise',
+                'Lasso',
+                'WeightDecay'])
