+++ chainer/optimizer_hooks/__init__.py
@@ -1,0 +1,5 @@
+from chainer.optimizer_hooks.gradient_clipping import GradientClipping  # NOQA
+from chainer.optimizer_hooks.gradient_hard_clipping import GradientHardClipping  # NOQA
+from chainer.optimizer_hooks.gradient_noise import GradientNoise  # NOQA
+from chainer.optimizer_hooks.lasso import Lasso  # NOQA
+from chainer.optimizer_hooks.weight_decay import WeightDecay  # NOQA
