+++ chainer/optimizer_hooks/gradient_clipping.py
@@ -1,0 +1,45 @@
+import collections
+
+import numpy
+import six
+
+from chainer import cuda
+
+
+def _sum_sqnorm(arr):
+    sq_sum = collections.defaultdict(float)
+    for x in arr:
+        with cuda.get_device_from_array(x) as dev:
+            x = x.ravel()
+            s = x.dot(x)
+            sq_sum[int(dev)] += s
+    return sum([float(i) for i in six.itervalues(sq_sum)])
+
+
+class GradientClipping(object):
+    """Optimizer hook function for gradient clipping.
+
+    This hook function scales all gradient arrays to fit to the defined L2 norm
+    threshold.
+
+    Args:
+        threshold (float): L2 norm threshold.
+
+    Attributes:
+        threshold (float): L2 norm threshold of gradient norm.
+
+    """
+    name = 'GradientClipping'
+
+    def __init__(self, threshold):
+        self.threshold = threshold
+
+    def __call__(self, opt):
+        norm = numpy.sqrt(_sum_sqnorm(
+            [p.grad for p in opt.target.params(False)]))
+        rate = self.threshold / norm
+        if rate < 1:
+            for param in opt.target.params(False):
+                grad = param.grad
+                with cuda.get_device_from_array(grad):
+                    grad *= rate
