+++ docs/source/reference/optimizer_hooks.rst
@@ -1,0 +1,14 @@
+Optimizer hook functions
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. module:: chainer.optimizer_hooks
+
+.. autosummary::
+   :toctree: generated/
+   :nosignatures:
+
+   chainer.optimizer_hooks.WeightDecay
+   chainer.optimizer_hooks.Lasso
+   chainer.optimizer_hooks.GradientClipping
+   chainer.optimizer_hooks.GradientHardClipping
+   chainer.optimizer_hooks.GradientNoise
